Add files via upload
I created a Music Playlist Manager in Python to offer a comprehensive solution for organizing and enjoying music collections. The system provides an array of essential features, including the addition and deletion of songs, alphabetical sorting for playlist organization, and the ability to maintain a list of recently played tracks. These functionalities ensure a versatile and user-friendly experience. The project also leverages a linked list data structure for efficient song storage and retrieval. This Music Playlist Manager project not only highlights my proficiency in Python programming but also underscores my dedication to creating software that enhances user convenience and interaction. By including this project in my portfolio and on GitHub, I aim to demonstrate my ability to design and develop software solutions that cater to real-world needs while showcasing strong problem-solving skills and a commitment to enhancing the user's daily life.